requirements.txt
tensorflow>=2.0.0
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0 